some merge errors and restructuring
