+++ client/src/components/CollagePreview.jsx
