+++ client/src/components/VisitComponent.jsx
@@ -22,17 +22,10 @@ function VisitComponent({ headingText, contentArray }) {
 				onMouseLeave={() => setIsActive(false)}
 				onTouchStart={() => (isActive ? setIsActive(false) : setIsActive(true))}
 			>
-				<Box
-					width="100%"
-					height="100%"
-					borderRadius="sm"
-					textStyle="battambang"
-				>
-					{
-                        contentArray.map((contentLine, index) => {
-                            return <VisitTextItem contentText={contentLine} />
-						})
-					}
+				<Box width="100%" height="100%" borderRadius="sm" textStyle="battambang">
+					{contentArray.map((contentLine, index) => {
+						return <VisitTextItem key={index} contentText={contentLine} />;
+					})}
 				</Box>
 			</Box>
 		</Box>
