+++ client/src/sections/Featured.jsx
@@ -8,55 +8,55 @@ function Featured() {
 	const [duchampImagesInfo, setDuchampImagesInfo] = useState([]);
 	const randomImageArray = [];
 
-	const fillDuchampImages = async () => {
-		const response = await fetch("/api/aic/marcel-duchamp", {
-			method: "GET",
-		});
-		const data = await response.json();
-		console.log(data);
-		setDuchampImagesData(data);
-	};
+	// const fillDuchampImages = async () => {
+	// 	const response = await fetch("/api/aic/marcel-duchamp", {
+	// 		method: "GET",
+	// 	});
+	// 	const data = await response.json();
+	// 	console.log(data);
+	// 	setDuchampImagesData(data);
+	// };
 
-	const selectRandomArtwork = async () => {
-		if (Object.keys(duchampImagesData).length > 0) {
-			const randomIndex = Math.floor(Math.random() * duchampImagesData.data.length);
-			const randomArtwork = duchampImagesData.data[randomIndex];
-			if (!randomImageArray.includes(randomArtwork)) {
-				randomImageArray.push(randomArtwork);
-				const randomArtWorkImageURL =
-					duchampImagesData.config.iiif_url +
-					"/" +
-					randomArtwork.image_id +
-					"/full/843,/0/default.jpg";
-				console.log(randomArtwork);
-				setDuchampImagesInfo((duchampImagesInfo) => [
-					...duchampImagesInfo,
-					{
-						url: randomArtWorkImageURL,
-						title: randomArtwork.title,
-						date: randomArtwork.date_display,
-						alt: randomArtwork.thumbnail.alt_text,
-					},
-				]);
-			} else {
-				selectRandomArtwork();
-			}
-		}
-	};
+	// const selectRandomArtwork = async () => {
+	// 	if (Object.keys(duchampImagesData).length > 0) {
+	// 		const randomIndex = Math.floor(Math.random() * duchampImagesData.data.length);
+	// 		const randomArtwork = duchampImagesData.data[randomIndex];
+	// 		if (!randomImageArray.includes(randomArtwork)) {
+	// 			randomImageArray.push(randomArtwork);
+	// 			const randomArtWorkImageURL =
+	// 				duchampImagesData.config.iiif_url +
+	// 				"/" +
+	// 				randomArtwork.image_id +
+	// 				"/full/843,/0/default.jpg";
+	// 			console.log(randomArtwork);
+	// 			setDuchampImagesInfo((duchampImagesInfo) => [
+	// 				...duchampImagesInfo,
+	// 				{
+	// 					url: randomArtWorkImageURL,
+	// 					title: randomArtwork.title,
+	// 					date: randomArtwork.date_display,
+	// 					alt: randomArtwork.thumbnail.alt_text,
+	// 				},
+	// 			]);
+	// 		} else {
+	// 			selectRandomArtwork();
+	// 		}
+	// 	}
+	// };
 
-	useEffect(() => {
-		fillDuchampImages();
-	}, []);
+	// useEffect(() => {
+	// 	fillDuchampImages();
+	// }, []);
 
-	// random math to pull a random artwork from the array
-	useEffect(() => {
-		selectRandomArtwork();
-		selectRandomArtwork();
-		selectRandomArtwork();
-		selectRandomArtwork();
-		selectRandomArtwork();
-		selectRandomArtwork();
-	}, [duchampImagesData]);
+	// // random math to pull a random artwork from the array
+	// useEffect(() => {
+	// 	selectRandomArtwork();
+	// 	selectRandomArtwork();
+	// 	selectRandomArtwork();
+	// 	selectRandomArtwork();
+	// 	selectRandomArtwork();
+	// 	selectRandomArtwork();
+	// }, [duchampImagesData]);
 
 	// async function getMetTestResponse() {
 	// 	const response = await fetch("/api/met/", {
@@ -212,14 +212,14 @@ function Featured() {
 						width={{ base: "330%", md: "270%", lg: "auto" }}
 						justify="space-around"
 					>
-						<FeaturedImage
+						{/* <FeaturedImage
 							flexDirection={{ base: "column-reverse", lg: "row-reverse" }}
 							alignment={{ base: "flex-start", lg: "flex-end" }}
 							title={duchampImagesInfo[0].title}
 							date={duchampImagesInfo[0].date_display}
 							url={duchampImagesInfo[0].url}
 							alt={duchampImagesInfo[0].alt_text}
-						/>
+						/> */}
 						<FeaturedImage
 							flexDirection={{ base: "column", lg: "row" }}
 							alignment={{ base: "flex-end", lg: "flex-start" }}
